% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{d_sp_map}
\alias{d_sp_map}
\title{Census map from Sao Paulo Municipality}
\format{A data frame with 847,280 rows and 7 variables:
\describe{
  \item{lat}{latitude}
  \item{long}{longitude}
  \item{order}{maptools stuff. This is not important}
  \item{hole}{maptools stuff. This is not important}
  \item{piece}{maptools stuff. This is not important}
  \item{group}{maptools stuff. This is not important}
  \item{id}{maptools stuff. This is not important}
}}
\source{
\url{https://www.ibge.gov.br/}
}
\usage{
d_sp_map
}
\description{
A dataset containing coordinates to make censitary regions map
}
\keyword{datasets}
