% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{educ}
\alias{educ}
\title{Censitary literacy information}
\format{A data frame with 18,363 rows and 6 variables:
\describe{
  \item{cod}{censitary region code}
  \item{situacao}{censitary code status}
  \item{alfab5}{literate population size (above five years old)}
  \item{total5}{population size (above five years old)}
  \item{prop}{alfab5/total5}
  \item{prop_cat}{maptools stuff. This is not important}
}}
\source{
\url{https://www.ibge.gov.br/}
}
\usage{
educ
}
\description{
A dataset containing censitary codes and literacy information
}
\keyword{datasets}
