% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{partes}
\alias{partes}
\title{Litigation parties}
\format{A data frame with 8,584,235 rows and 4 variables:
\describe{
  \item{id}{litigation id (internal)}
  \item{id_pessoa}{person id}
  \item{tipo_parte}{person side (plaintiff = autor, defendant = reu)}
  \item{endereco}{person address}
}}
\source{
\url{http://www.tjsp.jus.br/}
}
\usage{
partes
}
\description{
A dataset containing information about the parties of \code{processos}
}
\keyword{datasets}
