% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{processos}
\alias{processos}
\title{Litigation data}
\format{A data frame with 227,664 rows and 12 variables:
\describe{
  \item{id}{litigation id (internal)}
  \item{num}{litigation id (CNJ number, external)}
  \item{num_dep}{cases related to the litigation}
  \item{vara}{name of the court}
  \item{juiz}{judge name}
  \item{classificacao}{litigation type}
  \item{classe}{TPU Class from CNJ Res. 46}
  \item{assunto}{TPU Subject from CNJ Res. 46}
  \item{vl_causa}{case start value}
  \item{dt_distribuicao}{litigation start date}
  \item{dt_mov}{last movement date}
  \item{tempo}{time between start and last movement, in days}
}}
\source{
\url{http://www.tjsp.jus.br/}
}
\usage{
processos
}
\description{
A dataset containing information about litigation received from TJSP
}
\keyword{datasets}
